# docs/context/engineering/api-contracts/v1/openapi.yaml
openapi: 3.1.0
info:
  title: InComedy Server API
  version: "1.0.0"
  description: |
    Current backend API contract for auth, identity foundation, organizer workspace foundation,
    and service health endpoints.
servers:
  - url: https://incomedy.ru
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/auth/vk/start:
    get:
      tags: [Auth]
      summary: Issue a VK ID browser launch URL
      operationId: getAuthVkStart
      parameters:
        - in: header
          name: X-Request-ID
          required: false
          description: Optional client-provided UUID request id for tracing.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: VK ID launch URL returned successfully
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VkIdStartResponse"
        "429":
          description: Too many requests
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: VK auth is not configured in the current runtime
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/vk/verify:
    post:
      tags: [Auth]
      summary: Verify VK ID callback data and issue an internal session
      operationId: postAuthVkVerify
      parameters:
        - in: header
          name: X-Request-ID
          required: false
          description: Optional client-provided UUID request id for tracing.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VkIdVerifyPayload"
      responses:
        "200":
          description: VK auth completed successfully
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          description: Invalid payload or invalid signed auth state
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: VK code exchange or VK user info retrieval failed
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body exceeds route size limit
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: VK auth is not configured in the current runtime
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    VkIdStartResponse:
      type: object
      additionalProperties: false
      required:
        - auth_url
        - state
      properties:
        auth_url:
          type: string
          format: uri
        state:
          type: string
    VkIdVerifyPayload:
      type: object
      additionalProperties: false
      required:
        - code
        - state
        - device_id
      properties:
        code:
          type: string
        state:
          type: string
        device_id:
          type: string
    AuthSessionResponse:
      type: object
      additionalProperties: false
      required:
        - provider
        - access_token
        - refresh_token
        - expires_in
        - user
      properties:
        provider:
          type: string
          enum: [password, phone, vk, telegram, google, apple]
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
          example: 3600
        user:
          $ref: "#/components/schemas/SessionUser"
...
    SessionUser:
      type: object
      additionalProperties: false
      required:
        - id
        - display_name
        - roles
        - linked_providers
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
          example: John Doe
        username:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        roles:
          type: array
          items:
            type: string
            enum: [audience, comedian, organizer]
        active_role:
          type: string
          nullable: true
          enum: [audience, comedian, organizer]
        linked_providers:
          type: array
          items:
            type: string
            enum: [password, phone, vk, telegram, google, apple]
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: auth_failed
        message:
          type: string
          example: Authentication failed